move dbaccessors to warehouse app
